NASAEARTH-5 refactoring to simplify dataset-tab
