+++ src/components/App.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { App } from "./App";
+import { App } from "./app";
 import { render, screen } from "@testing-library/react";
 
 describe("test load app", () => {
