+++ src/utils/embed-check.ts
@@ -1,0 +1,8 @@
+/**
+ * Checks if the application is running in non-embedded mode
+ * @returns true if the noEmbed parameter is present in the URL
+ */
+export const isNonEmbedded = () => {
+  const urlParams = new URLSearchParams(window.location.search);
+  return urlParams.has("noEmbed");
+};
